+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -262,6 +262,9 @@ class SentryClientTest {
         assertEquals("fp", event.fingerprints[0])
         assertEquals("id", event.user.id)
         assertEquals(SentryLevel.FATAL, event.level)
+        assertNotNull(event.request) {
+            assertEquals("post", it.method)
+        }
     }
 
     @Test
@@ -755,6 +758,25 @@ class SentryClientTest {
     }
 
     @Test
+    fun `when captureTransactions unfinished spans are removed`() {
+        val sut = fixture.getSut()
+        val transaction = SentryTransaction("a-transaction", "op")
+        val span1 = transaction.startChild("span1")
+        span1.finish()
+        val span2 = transaction.startChild("span2")
+
+        sut.captureTransaction(transaction, mock(), null)
+        verify(fixture.transport).send(check {
+            val sentTransaction = it.items.first().getTransaction(fixture.sentryOptions.serializer)
+                assertNotNull(sentTransaction) { tx ->
+                    val sentSpanIds = tx.spans.map { span -> span.spanId }
+                    assertTrue(sentSpanIds.contains(span1.spanContext.spanId))
+                    assertFalse(sentSpanIds.contains(span2.spanContext.spanId))
+                }
+        }, eq(null))
+    }
+
+    @Test
     fun `when captureTransaction with attachments`() {
         val transaction = SentryTransaction("a-transaction", "op")
         fixture.getSut().captureTransaction(transaction, createScopeWithAttachments(), null)
@@ -854,6 +876,9 @@ class SentryClientTest {
             level = SentryLevel.FATAL
             user = User().apply {
                 id = "id"
+            }
+            request = Request().apply {
+                method = "post"
             }
         }
     }